+++ ndexbiogridloader/__init__.py
@@ -3,5 +3,5 @@
 """Top-level package for NDEx BioGRID Content Loader."""
 
 __author__ = """Vladimir Rynkov"""
-__email__ = 'vrynkov@ucsd.edu'
-__version__ = '0.1.0'
+__email__ = 'vrynkov@yahoo.com'
+__version__ = '0.2.0'