+++ HISTORY.rst
@@ -2,6 +2,14 @@
 History
 =======
 
+0.2.0 (2019-07-26)
+------------------
+
+* Collapsed duplicate edges.
+* Added a new optional argument skipdownload to command line.
+* Added integration tests.
+
+
 0.1.0 (2019-05-10)
 ------------------